==> 1-split-data.ipynb <==
import pandas as pd
from sklearn.model_selection import train_test_split

df = pd.read_csv('PURCHASE ORDER DATA EXTRACT 2012-2015_0.csv')

train_df, test_df = train_test_split(df, test_size=1000, random_state=42)
train_df = train_df.sort_index()
test_df = test_df.sort_index()
print(train_df.shape)
print(test_df.shape)

train_df.to_csv('train.csv', sep=',')
test_df.to_csv('test.csv', sep=',')
