==> 3-verify.ipynb <==
import pandas as pd
import os
import numpy as np
from tqdm.notebook import tqdm

df = pd.read_csv('train.csv', index_col=0)

total_tasks = 3460
embedding_indices = set()

for i in range(total_tasks):
    emb_f = f'training_embeddings/training_embeddings_{i}.npy'
    if os.path.isfile(emb_f):
        embedding = np.load(emb_f)
        for j in embedding[:,0]:
            embedding_indices.add(j)
    else:
        print(f'Missing embedding file: {emb_f}')

print(f'Found {len(embedding_indices)} embeddings')

for i in df.index:
    if i not in embedding_indices:
        print(f'Missing embedding {i}')
