==> 4-neural-network.ipynb <==
import torch
import os
import numpy as np
import pandas as pd
import math
from tqdm.notebook import tqdm
from torch.utils.data import Dataset, DataLoader
import torch.nn.functional as Fun
from utils import price_to_log_cat, get_date_to_month_buckets, quantity_to_log_cat

# total_tasks = 3460
total_tasks = 3460
per_task = 100
embeddings = np.empty(0)
embedding_indices = set()

for i in tqdm(range(total_tasks)):
    emb_f = f'training_embeddings/training_embeddings_{i}.npy'
    if os.path.isfile(emb_f):
        embedding = np.load(emb_f)
        if embeddings.size == 0:
            embeddings = embedding[:,1:]
        else:
            embeddings = np.vstack((embeddings, embedding[:,1:]))
    else:
        print(f'Missing embedding file: {emb_f}')

print(f'Found {len(embedding_indices)} embeddings: {embeddings.shape}')
train_df = pd.read_csv('train.csv', index_col=0)
train_df.shape

train_df['Total Price'] = train_df['Total Price'].apply(price_to_log_cat).to_numpy()
train_df['Unit Price'] = train_df['Unit Price'].apply(price_to_log_cat).to_numpy()

# Improve date sanitization with GPT cleaning
train_df['Purchase Date'] = pd.to_datetime([date[:-4]+'20'+date[-2:] if type(date) == str else date for date in train_df['Purchase Date'].values], format='%m/%d/%Y')
train_df['Creation Date'] = pd.to_datetime(train_df['Creation Date'], format='%m/%d/%Y')
date_to_month_buckets = get_date_to_month_buckets(train_df['Creation Date'].min())
train_df['Creation Date'] = train_df['Creation Date'].apply(date_to_month_buckets).to_numpy()
train_df['Purchase Date'] = train_df['Purchase Date'].apply(date_to_month_buckets).to_numpy()

train_df['Quantity'] = train_df['Quantity'].apply(quantity_to_log_cat).to_numpy()

class Task_Dataset(Dataset):
    def __init__(self, X : np.ndarray, 
                       y : np.ndarray):
        self.X = X
        self.y = Fun.one_hot(torch.from_numpy(y).to(torch.int64))
        assert self.X.shape[0] == self.y.shape[0]
        
    def __len__(self):
        return len(self.y)
    
    def __getitem__(self, idx):    
        X = torch.from_numpy(self.X[idx,:]).float().squeeze()
        y = self.y[idx]
        return X, y

total_price_ds = Task_Dataset(embeddings, train_df.iloc[:total_tasks*per_task]['Total Price'].to_numpy())
total_price_dl = DataLoader(total_price_ds, batch_size = 64, shuffle = True)

unit_price_ds = Task_Dataset(embeddings, train_df.iloc[:total_tasks*per_task]['Unit Price'].to_numpy())
unit_price_dl = DataLoader(unit_price_ds, batch_size = 64, shuffle = True)

creation_date_ds = Task_Dataset(embeddings, train_df.iloc[:total_tasks*per_task]['Creation Date'].to_numpy())
creation_date_dl = DataLoader(creation_date_ds, batch_size = 64, shuffle = True)

purchase_date_ds = Task_Dataset(embeddings, train_df.iloc[:total_tasks*per_task]['Purchase Date'].to_numpy())
purchase_date_dl = DataLoader(purchase_date_ds, batch_size = 64, shuffle = True)

quantity_ds = Task_Dataset(embeddings, train_df.iloc[:total_tasks*per_task]['Quantity'].to_numpy())
quantity_dl = DataLoader(quantity_ds, batch_size = 64, shuffle = True)

class MultiTask_Network(torch.nn.Module):
    def __init__(self, input_dim, 
                 output_dim_0 : int = 1,
                 output_dim_1 : int = 1,
                 output_dim_2 : int = 1,
                 output_dim_3 : int = 1,
                 output_dim_4 : int = 1,
                 hidden_dim : int = 2048):
        
        super(MultiTask_Network, self).__init__()
        self.input_dim = input_dim
        self.output_dim_0 = output_dim_0
        self.output_dim_1 = output_dim_1
        self.output_dim_2 = output_dim_2
        self.output_dim_3 = output_dim_3
        self.output_dim_4 = output_dim_4
        self.hidden_dim = hidden_dim
        
        self.hidden0 = torch.nn.Linear(self.input_dim, self.hidden_dim)
        self.hidden1 = torch.nn.Linear(self.hidden_dim, self.hidden_dim)
        self.hidden2 = torch.nn.Linear(self.hidden_dim, self.hidden_dim)
        self.final_0 = torch.nn.Linear(self.hidden_dim, self.output_dim_0)
        self.final_1 = torch.nn.Linear(self.hidden_dim, self.output_dim_1)
        self.final_2 = torch.nn.Linear(self.hidden_dim, self.output_dim_2)
        self.final_3 = torch.nn.Linear(self.hidden_dim, self.output_dim_3)
        self.final_4 = torch.nn.Linear(self.hidden_dim, self.output_dim_4)
        
    def forward(self, x : torch.Tensor, task_id : int):
        x = self.hidden0(x)
        x = torch.relu(x)
        x = self.hidden1(x)
        x = torch.relu(x)
        x = self.hidden2(x)
        x = torch.relu(x)
        if task_id == 0:
            x = self.final_0(x)
        elif task_id == 1:
            x = self.final_1(x)
        elif task_id == 2:
            x = self.final_2(x)
        elif task_id == 3:
            x = self.final_3(x)
        elif task_id == 4:
            x = self.final_4(x)
        else:
            assert False, 'Bad Task ID passed'
        
        return x

model = MultiTask_Network(total_price_ds.X.shape[1],
                          output_dim_0 = total_price_ds.y.shape[1],
                          output_dim_1 = unit_price_ds.y.shape[1],
                          output_dim_2 = creation_date_ds.y.shape[1],
                          output_dim_3 = purchase_date_ds.y.shape[1],
                          output_dim_4 = quantity_ds.y.shape[1])
optimizer = torch.optim.Adam(model.parameters(), lr = 1e-4)
loss_fn_0 = torch.nn.CrossEntropyLoss()
loss_fn_1 = torch.nn.CrossEntropyLoss()
loss_fn_2 = torch.nn.CrossEntropyLoss()
loss_fn_3 = torch.nn.CrossEntropyLoss()
loss_fn_4 = torch.nn.CrossEntropyLoss()

print(total_price_ds.X.shape[1],
                          total_price_ds.y.shape[1],
                          unit_price_ds.y.shape[1],
                          creation_date_ds.y.shape[1],
                          purchase_date_ds.y.shape[1],
                          quantity_ds.y.shape[1])

for i in range(10):
    zipped_dls = zip(total_price_dl, unit_price_dl, creation_date_dl, purchase_date_dl, quantity_dl)
    for j, ((batch_X_0, batch_y_0), (batch_X_1, batch_y_1), (batch_X_2, batch_y_2), (batch_X_3, batch_y_3), (batch_X_4, batch_y_4)) in (pbar := tqdm(enumerate(zipped_dls), unit="batch", total=len(total_price_dl))):
        preds_0 = model(batch_X_0, task_id=0)
        loss_0 = loss_fn_0(preds_0, batch_y_0.float())
        preds_1 = model(batch_X_1, task_id=1)
        loss_1 = loss_fn_1(preds_1, batch_y_1.float())
        preds_2 = model(batch_X_2, task_id=2)
        loss_2 = loss_fn_2(preds_2, batch_y_2.float())
        preds_3 = model(batch_X_3, task_id=3)
        loss_3 = loss_fn_3(preds_3, batch_y_3.float())
        preds_4 = model(batch_X_4, task_id=4)
        loss_4 = loss_fn_4(preds_4, batch_y_4.float())
        loss = loss_0 + loss_1 + loss_2 + loss_3 + loss_4
        pbar.set_description("Loss = " + f"{loss:.4f}"[:6])
        
        optimizer.zero_grad()
        loss.backward()
        optimizer.step()
    n_correct = [0]*5
    n_test = [0]*5
    zipped_dls = zip(total_price_dl, unit_price_dl, creation_date_dl, purchase_date_dl, quantity_dl)
    for j, ((batch_X_0, batch_y_0), (batch_X_1, batch_y_1), (batch_X_2, batch_y_2), (batch_X_3, batch_y_3), (batch_X_4, batch_y_4)) in (pbar := tqdm(enumerate(zipped_dls), unit="batch", total=len(total_price_dl))):
        preds_0 = model(batch_X_0, task_id=0)
        preds_1 = model(batch_X_1, task_id=1)
        preds_2 = model(batch_X_2, task_id=2)
        preds_3 = model(batch_X_3, task_id=3)
        preds_4 = model(batch_X_4, task_id=4)
        n_correct[0] += (preds_0.detach().numpy().argmax(axis=1) == batch_y_0.numpy().argmax(axis=1)).sum()
        n_correct[1] += (preds_1.detach().numpy().argmax(axis=1) == batch_y_1.numpy().argmax(axis=1)).sum()
        n_correct[2] += (preds_2.detach().numpy().argmax(axis=1) == batch_y_2.numpy().argmax(axis=1)).sum()
        n_correct[3] += (preds_3.detach().numpy().argmax(axis=1) == batch_y_3.numpy().argmax(axis=1)).sum()
        n_correct[4] += (preds_4.detach().numpy().argmax(axis=1) == batch_y_4.numpy().argmax(axis=1)).sum()
        n_test[0] += preds_0.shape[0]
        n_test[1] += preds_1.shape[0]
        n_test[2] += preds_2.shape[0]
        n_test[3] += preds_3.shape[0]
        n_test[4] += preds_4.shape[0]
    print(f'Epoch {i} accuracy: total_price={n_correct[0]/n_test[0]*100} unit_price={n_correct[1]/n_test[1]*100} creation_date={n_correct[2]/n_test[2]*100} purchase_date={n_correct[3]/n_test[3]*100} quantity_dl={n_correct[4]/n_test[4]*100}')


for i in range(5):
    zipped_dls = zip(total_price_dl, unit_price_dl, creation_date_dl, purchase_date_dl, quantity_dl)
    for j, ((batch_X_0, batch_y_0), (batch_X_1, batch_y_1), (batch_X_2, batch_y_2), (batch_X_3, batch_y_3), (batch_X_4, batch_y_4)) in (pbar := tqdm(enumerate(zipped_dls), unit="batch", total=len(total_price_dl))):
        preds_0 = model(batch_X_0, task_id=0)
        loss_0 = loss_fn_0(preds_0, batch_y_0.float())
        preds_1 = model(batch_X_1, task_id=1)
        loss_1 = loss_fn_1(preds_1, batch_y_1.float())
        preds_2 = model(batch_X_2, task_id=2)
        loss_2 = loss_fn_2(preds_2, batch_y_2.float())
        preds_3 = model(batch_X_3, task_id=3)
        loss_3 = loss_fn_3(preds_3, batch_y_3.float())
        preds_4 = model(batch_X_4, task_id=4)
        loss_4 = loss_fn_4(preds_4, batch_y_4.float())
        loss = loss_0 + loss_1 + loss_2 + loss_3 + loss_4
        pbar.set_description("Loss = " + f"{loss:.4f}"[:6])
        
        optimizer.zero_grad()
        loss.backward()
        optimizer.step()
    n_correct = [0]*5
    n_test = [0]*5
    zipped_dls = zip(total_price_dl, unit_price_dl, creation_date_dl, purchase_date_dl, quantity_dl)
    for j, ((batch_X_0, batch_y_0), (batch_X_1, batch_y_1), (batch_X_2, batch_y_2), (batch_X_3, batch_y_3), (batch_X_4, batch_y_4)) in (pbar := tqdm(enumerate(zipped_dls), unit="batch", total=len(total_price_dl))):
        preds_0 = model(batch_X_0, task_id=0)
        preds_1 = model(batch_X_1, task_id=1)
        preds_2 = model(batch_X_2, task_id=2)
        preds_3 = model(batch_X_3, task_id=3)
        preds_4 = model(batch_X_4, task_id=4)
        n_correct[0] += (preds_0.detach().numpy().argmax(axis=1) == batch_y_0.numpy().argmax(axis=1)).sum()
        n_correct[1] += (preds_1.detach().numpy().argmax(axis=1) == batch_y_1.numpy().argmax(axis=1)).sum()
        n_correct[2] += (preds_2.detach().numpy().argmax(axis=1) == batch_y_2.numpy().argmax(axis=1)).sum()
        n_correct[3] += (preds_3.detach().numpy().argmax(axis=1) == batch_y_3.numpy().argmax(axis=1)).sum()
        n_correct[4] += (preds_4.detach().numpy().argmax(axis=1) == batch_y_4.numpy().argmax(axis=1)).sum()
        n_test[0] += preds_0.shape[0]
        n_test[1] += preds_1.shape[0]
        n_test[2] += preds_2.shape[0]
        n_test[3] += preds_3.shape[0]
        n_test[4] += preds_4.shape[0]
    print(f'Epoch {i} accuracy: total_price={n_correct[0]/n_test[0]*100} unit_price={n_correct[1]/n_test[1]*100} creation_date={n_correct[2]/n_test[2]*100} purchase_date={n_correct[3]/n_test[3]*100} quantity_dl={n_correct[4]/n_test[4]*100}')


for i in range(5):
    zipped_dls = zip(total_price_dl, unit_price_dl, creation_date_dl, purchase_date_dl, quantity_dl)
    for j, ((batch_X_0, batch_y_0), (batch_X_1, batch_y_1), (batch_X_2, batch_y_2), (batch_X_3, batch_y_3), (batch_X_4, batch_y_4)) in (pbar := tqdm(enumerate(zipped_dls), unit="batch", total=len(total_price_dl))):
        preds_0 = model(batch_X_0, task_id=0)
        loss_0 = loss_fn_0(preds_0, batch_y_0.float())
        preds_1 = model(batch_X_1, task_id=1)
        loss_1 = loss_fn_1(preds_1, batch_y_1.float())
        preds_2 = model(batch_X_2, task_id=2)
        loss_2 = loss_fn_2(preds_2, batch_y_2.float())
        preds_3 = model(batch_X_3, task_id=3)
        loss_3 = loss_fn_3(preds_3, batch_y_3.float())
        preds_4 = model(batch_X_4, task_id=4)
        loss_4 = loss_fn_4(preds_4, batch_y_4.float())
        loss = loss_0 + loss_1 + loss_2 + loss_3 + loss_4
        pbar.set_description("Loss = " + f"{loss:.4f}"[:6])
        
        optimizer.zero_grad()
        loss.backward()
        optimizer.step()
    n_correct = [0]*5
    n_test = [0]*5
    zipped_dls = zip(total_price_dl, unit_price_dl, creation_date_dl, purchase_date_dl, quantity_dl)
    for j, ((batch_X_0, batch_y_0), (batch_X_1, batch_y_1), (batch_X_2, batch_y_2), (batch_X_3, batch_y_3), (batch_X_4, batch_y_4)) in (pbar := tqdm(enumerate(zipped_dls), unit="batch", total=len(total_price_dl))):
        preds_0 = model(batch_X_0, task_id=0)
        preds_1 = model(batch_X_1, task_id=1)
        preds_2 = model(batch_X_2, task_id=2)
        preds_3 = model(batch_X_3, task_id=3)
        preds_4 = model(batch_X_4, task_id=4)
        n_correct[0] += (preds_0.detach().numpy().argmax(axis=1) == batch_y_0.numpy().argmax(axis=1)).sum()
        n_correct[1] += (preds_1.detach().numpy().argmax(axis=1) == batch_y_1.numpy().argmax(axis=1)).sum()
        n_correct[2] += (preds_2.detach().numpy().argmax(axis=1) == batch_y_2.numpy().argmax(axis=1)).sum()
        n_correct[3] += (preds_3.detach().numpy().argmax(axis=1) == batch_y_3.numpy().argmax(axis=1)).sum()
        n_correct[4] += (preds_4.detach().numpy().argmax(axis=1) == batch_y_4.numpy().argmax(axis=1)).sum()
        n_test[0] += preds_0.shape[0]
        n_test[1] += preds_1.shape[0]
        n_test[2] += preds_2.shape[0]
        n_test[3] += preds_3.shape[0]
        n_test[4] += preds_4.shape[0]
    print(f'Epoch {i} accuracy: total_price={n_correct[0]/n_test[0]*100} unit_price={n_correct[1]/n_test[1]*100} creation_date={n_correct[2]/n_test[2]*100} purchase_date={n_correct[3]/n_test[3]*100} quantity_dl={n_correct[4]/n_test[4]*100}')


for i in range(10):
    zipped_dls = zip(total_price_dl, unit_price_dl, creation_date_dl, purchase_date_dl, quantity_dl)
    for j, ((batch_X_0, batch_y_0), (batch_X_1, batch_y_1), (batch_X_2, batch_y_2), (batch_X_3, batch_y_3), (batch_X_4, batch_y_4)) in (pbar := tqdm(enumerate(zipped_dls), unit="batch", total=len(total_price_dl))):
        preds_0 = model(batch_X_0, task_id=0)
        loss_0 = loss_fn_0(preds_0, batch_y_0.float())
        preds_1 = model(batch_X_1, task_id=1)
        loss_1 = loss_fn_1(preds_1, batch_y_1.float())
        preds_2 = model(batch_X_2, task_id=2)
        loss_2 = loss_fn_2(preds_2, batch_y_2.float())
        preds_3 = model(batch_X_3, task_id=3)
        loss_3 = loss_fn_3(preds_3, batch_y_3.float())
        preds_4 = model(batch_X_4, task_id=4)
        loss_4 = loss_fn_4(preds_4, batch_y_4.float())
        loss = loss_0 + loss_1 + loss_2 + loss_3 + loss_4
        pbar.set_description("Loss = " + f"{loss:.4f}"[:6])
        
        optimizer.zero_grad()
        loss.backward()
        optimizer.step()
    n_correct = [0]*5
    n_test = [0]*5
    zipped_dls = zip(total_price_dl, unit_price_dl, creation_date_dl, purchase_date_dl, quantity_dl)
    for j, ((batch_X_0, batch_y_0), (batch_X_1, batch_y_1), (batch_X_2, batch_y_2), (batch_X_3, batch_y_3), (batch_X_4, batch_y_4)) in (pbar := tqdm(enumerate(zipped_dls), unit="batch", total=len(total_price_dl))):
        preds_0 = model(batch_X_0, task_id=0)
        preds_1 = model(batch_X_1, task_id=1)
        preds_2 = model(batch_X_2, task_id=2)
        preds_3 = model(batch_X_3, task_id=3)
        preds_4 = model(batch_X_4, task_id=4)
        n_correct[0] += (preds_0.detach().numpy().argmax(axis=1) == batch_y_0.numpy().argmax(axis=1)).sum()
        n_correct[1] += (preds_1.detach().numpy().argmax(axis=1) == batch_y_1.numpy().argmax(axis=1)).sum()
        n_correct[2] += (preds_2.detach().numpy().argmax(axis=1) == batch_y_2.numpy().argmax(axis=1)).sum()
        n_correct[3] += (preds_3.detach().numpy().argmax(axis=1) == batch_y_3.numpy().argmax(axis=1)).sum()
        n_correct[4] += (preds_4.detach().numpy().argmax(axis=1) == batch_y_4.numpy().argmax(axis=1)).sum()
        n_test[0] += preds_0.shape[0]
        n_test[1] += preds_1.shape[0]
        n_test[2] += preds_2.shape[0]
        n_test[3] += preds_3.shape[0]
        n_test[4] += preds_4.shape[0]
    print(f'Epoch {i} accuracy: total_price={n_correct[0]/n_test[0]*100} unit_price={n_correct[1]/n_test[1]*100} creation_date={n_correct[2]/n_test[2]*100} purchase_date={n_correct[3]/n_test[3]*100} quantity_dl={n_correct[4]/n_test[4]*100}')


for i in range(10):
    zipped_dls = zip(total_price_dl, unit_price_dl, creation_date_dl, purchase_date_dl, quantity_dl)
    for j, ((batch_X_0, batch_y_0), (batch_X_1, batch_y_1), (batch_X_2, batch_y_2), (batch_X_3, batch_y_3), (batch_X_4, batch_y_4)) in (pbar := tqdm(enumerate(zipped_dls), unit="batch", total=len(total_price_dl))):
        preds_0 = model(batch_X_0, task_id=0)
        loss_0 = loss_fn_0(preds_0, batch_y_0.float())
        preds_1 = model(batch_X_1, task_id=1)
        loss_1 = loss_fn_1(preds_1, batch_y_1.float())
        preds_2 = model(batch_X_2, task_id=2)
        loss_2 = loss_fn_2(preds_2, batch_y_2.float())
        preds_3 = model(batch_X_3, task_id=3)
        loss_3 = loss_fn_3(preds_3, batch_y_3.float())
        preds_4 = model(batch_X_4, task_id=4)
        loss_4 = loss_fn_4(preds_4, batch_y_4.float())
        loss = loss_0 + loss_1 + loss_2 + loss_3 + loss_4
        pbar.set_description("Loss = " + f"{loss:.4f}"[:6])
        
        optimizer.zero_grad()
        loss.backward()
        optimizer.step()
    n_correct = [0]*5
    n_test = [0]*5
    zipped_dls = zip(total_price_dl, unit_price_dl, creation_date_dl, purchase_date_dl, quantity_dl)
    for j, ((batch_X_0, batch_y_0), (batch_X_1, batch_y_1), (batch_X_2, batch_y_2), (batch_X_3, batch_y_3), (batch_X_4, batch_y_4)) in (pbar := tqdm(enumerate(zipped_dls), unit="batch", total=len(total_price_dl))):
        preds_0 = model(batch_X_0, task_id=0)
        preds_1 = model(batch_X_1, task_id=1)
        preds_2 = model(batch_X_2, task_id=2)
        preds_3 = model(batch_X_3, task_id=3)
        preds_4 = model(batch_X_4, task_id=4)
        n_correct[0] += (preds_0.detach().numpy().argmax(axis=1) == batch_y_0.numpy().argmax(axis=1)).sum()
        n_correct[1] += (preds_1.detach().numpy().argmax(axis=1) == batch_y_1.numpy().argmax(axis=1)).sum()
        n_correct[2] += (preds_2.detach().numpy().argmax(axis=1) == batch_y_2.numpy().argmax(axis=1)).sum()
        n_correct[3] += (preds_3.detach().numpy().argmax(axis=1) == batch_y_3.numpy().argmax(axis=1)).sum()
        n_correct[4] += (preds_4.detach().numpy().argmax(axis=1) == batch_y_4.numpy().argmax(axis=1)).sum()
        n_test[0] += preds_0.shape[0]
        n_test[1] += preds_1.shape[0]
        n_test[2] += preds_2.shape[0]
        n_test[3] += preds_3.shape[0]
        n_test[4] += preds_4.shape[0]
    print(f'Epoch {i} accuracy: total_price={n_correct[0]/n_test[0]*100} unit_price={n_correct[1]/n_test[1]*100} creation_date={n_correct[2]/n_test[2]*100} purchase_date={n_correct[3]/n_test[3]*100} quantity_dl={n_correct[4]/n_test[4]*100}')


torch.save(model.state_dict(), './ckpt')

for i in range(10):
    zipped_dls = zip(total_price_dl, unit_price_dl, creation_date_dl, purchase_date_dl, quantity_dl)
    for j, ((batch_X_0, batch_y_0), (batch_X_1, batch_y_1), (batch_X_2, batch_y_2), (batch_X_3, batch_y_3), (batch_X_4, batch_y_4)) in (pbar := tqdm(enumerate(zipped_dls), unit="batch", total=len(total_price_dl))):
        preds_0 = model(batch_X_0, task_id=0)
        loss_0 = loss_fn_0(preds_0, batch_y_0.float())
        preds_1 = model(batch_X_1, task_id=1)
        loss_1 = loss_fn_1(preds_1, batch_y_1.float())
        preds_2 = model(batch_X_2, task_id=2)
        loss_2 = loss_fn_2(preds_2, batch_y_2.float())
        preds_3 = model(batch_X_3, task_id=3)
        loss_3 = loss_fn_3(preds_3, batch_y_3.float())
        preds_4 = model(batch_X_4, task_id=4)
        loss_4 = loss_fn_4(preds_4, batch_y_4.float())
        loss = loss_0 + loss_1 + loss_2 + loss_3 + loss_4
        pbar.set_description("Loss = " + f"{loss:.4f}"[:6])
        
        optimizer.zero_grad()
        loss.backward()
        optimizer.step()
    n_correct = [0]*5
    n_test = [0]*5
    zipped_dls = zip(total_price_dl, unit_price_dl, creation_date_dl, purchase_date_dl, quantity_dl)
    for j, ((batch_X_0, batch_y_0), (batch_X_1, batch_y_1), (batch_X_2, batch_y_2), (batch_X_3, batch_y_3), (batch_X_4, batch_y_4)) in (pbar := tqdm(enumerate(zipped_dls), unit="batch", total=len(total_price_dl))):
        preds_0 = model(batch_X_0, task_id=0)
        preds_1 = model(batch_X_1, task_id=1)
        preds_2 = model(batch_X_2, task_id=2)
        preds_3 = model(batch_X_3, task_id=3)
        preds_4 = model(batch_X_4, task_id=4)
        n_correct[0] += (preds_0.detach().numpy().argmax(axis=1) == batch_y_0.numpy().argmax(axis=1)).sum()
        n_correct[1] += (preds_1.detach().numpy().argmax(axis=1) == batch_y_1.numpy().argmax(axis=1)).sum()
        n_correct[2] += (preds_2.detach().numpy().argmax(axis=1) == batch_y_2.numpy().argmax(axis=1)).sum()
        n_correct[3] += (preds_3.detach().numpy().argmax(axis=1) == batch_y_3.numpy().argmax(axis=1)).sum()
        n_correct[4] += (preds_4.detach().numpy().argmax(axis=1) == batch_y_4.numpy().argmax(axis=1)).sum()
        n_test[0] += preds_0.shape[0]
        n_test[1] += preds_1.shape[0]
        n_test[2] += preds_2.shape[0]
        n_test[3] += preds_3.shape[0]
        n_test[4] += preds_4.shape[0]
    print(f'Epoch {i} accuracy: total_price={n_correct[0]/n_test[0]*100} unit_price={n_correct[1]/n_test[1]*100} creation_date={n_correct[2]/n_test[2]*100} purchase_date={n_correct[3]/n_test[3]*100} quantity_dl={n_correct[4]/n_test[4]*100}')


torch.save(model.state_dict(), './ckpt2')
