==> 5-test-concealed.ipynb <==
import numpy as np
import pandas as pd

df = pd.read_csv('test.csv', index_col=0)

df.shape

np.random.seed(42)
concealed_index = np.random.choice(5, df.shape[0])

np.save('test_conceal_index.npy', concealed_index)
